+++ src/Exceptions/InvalidRequestException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenAI\Exceptions;
+
+final class InvalidRequestException extends ErrorException
+{
+}