+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- More specific exceptions: `InvalidApiKeyException` and `InvalidRequestException`
+
 ## v0.1.0 (2022-10-20)
 ### Added
 - First version